Naruto Base and 6th Path Model Rigged and Textured
They are in FBX format, so hopefully that is convenient. For the Asura 9 Tails Chakra, it still needs to be rigged, so I will spend some time working on that. I will try to keep the structure of the skeletons between each character as similar as possible, now that we have Naruto as a base character.